Set to log dimensions upon drawing
Obviously, by the time that they are drawn, they are loaded and have dimensions. Therefore, we can set the width and height again upon calling the draw function, which wouldn't really cause to much machine lag and would fix our problem. Later, when we are optimizing things, we'll come up with a more permanent solution.