Fixed arms not appearing
In order to do so, I had to change the order in which everything is
drawn, meaning that the arms are drawn on top of the torso. It doesn't
look as great, but the other way around doesn't work because then the
arms don't have a parent to attach themselves to. The way to rectify
this would be to draw the sprite twice per frame, potentially doubling
lag.