Moved a fair bit of code over to a new script,
which I named gameScript. As you might have already figured out by now,
gameScript is designed to hold all of the main scripts for the game,
whilst the other script holds all of the scripts for the menu and such.
I'll rename the script.js so that it actually and accurately describes
what it does. I also cleaned up and fixed the code now residing within
gameScript, as is described in the issue I recently closed. Next, I aim
to fix the wave animation, and after that I will create the tile system!
:D